♻️ refactor: test dir's unit&integration division
