+++ backend/tests/unit/modules/audio/test_audio.py
@@ -1,9 +1,7 @@
-import pytest
 from backend.app.modules.tts.utils import parse_tts_by_newlines
 
 
 
-# === Mocked ElevenLabs-like response ===
 FAKE_TTS_RESPONSE = {
     "audio_base64": "FAKE_AUDIO_DATA",
     "alignment": {
